added flags

+++ ShapeJumper.sh
@@ -1,4 +1,4 @@
-#!bin/bash
+#!/bin/bash
 #ShapeJumper.sh
 
 #June 29, 2020
@@ -6,56 +6,134 @@
 #Process all reads through ShapeJumper pipeline and outputs a list of all deletions
 #identified and their rates.
 
-#check for argument count
-if [ $# -eq 3 ]; then
-	fasta=$1
-	testFastq1=$2
-	controlFastq1=$3
-elif [ $# -eq 5 ]; then
-	fasta=$1
-	testFastq1=$2
-	testFastq2=$3
-	controlFastq1=$4
-	controlFastq2=$5
+set -e # exit on first error (if any)
 
-else
-	echo "Wrong number of arguments"
-	echo "USAGE: bash ShapeJumper.sh referenceSequence.fasta crosslinkRead1.fastq crosslinkread2.fastq controlRead1.fastq controlRead2.fastq"
+#handle arguments, including optional arguments
+script="ShapeJumper"
+#Declare the number of mandatory args
+margs=3
+
+#examples of usage
+function example {
+    echo -e "example usage - unpaired reads: \nbash ShapeJumper.sh -f referenceSequence.fasta -t1 crosslinkUnPairedReads.fastq -c1 controlUnPairedReads.fastq"
 	echo "OR"
-	echo "bash ShapeJumper.sh referenceSequence.fasta crosslinkUnPairedReads.fastq controlUnPairedReads.fastq"
-	exit 1
-fi
-	
-#check the files to make sure they exist
-if [ ! -e "$fasta" ]
-then
-	echo "ERROR: In $0, the input fasta file $fasta does not exist"
-	exit 1
-fi
+    echo -e "example usage - paired end reads: \nbash ShapeJumper.sh -f referenceSequence.fasta -t1 crosslinkRead1.fastq -t2 crosslinkread2.fastq -c1 controlRead1.fastq -c2 controlRead2.fastq"	
+}
 
-if [ ! -e "$testFastq1" ]
-then
-	echo "ERROR: In $0, the input fastq file $testFastq1 does not exist"
-	exit 1
-fi
+function usage {
+    echo -e "usage: bash $script -f FASTA -t1 TEST -c1 CONTROL [OPTIONALARGUMENTS]\n"
+}
 
-if [ ! -e "$controlFastq1" ]
-then
-	echo "ERROR: In $0, the input fastq file $controlFastq1 does not exist"
-	exit 1
-fi
+function help {
+  usage
+    echo -e "Mandatory Arguments:"
+    echo -e "  -f, --fasta  REFERENCEFASTA.FASTA  \n\"Fasta file containing sequences of all target RNAs.\"\n"
+    echo -e "  -t1, --test1  TESTREADS.FASTQ  \n\"Fastq file containing reads from test/crosslinked sample.\"\n"
+    echo -e "  -c1, --control1  CONTROLREADS.FASTQ  \n\"Fastq file containing reads from control/monoAdduct sample.\"\n"
+    echo -e "Optional Arguments:"
+    echo -e "  -t2, --test2   TESTREADS2.FASTQ   \n\"Fastq file containing paired reads 2 from test/crosslinked sample.\"\n"
+    echo -e "  -c2, --control2   CONTROLREADS2.FASTQ  \n\"Fastq file containing paired reads 2 from control/monoAdduct sample.\"\n"
+    echo -e "  -sc, --structureCassette \n\"This option renumbers the output deletions to remove structure cassettes.\"\n"
+    echo -e "  -s, --shift5Prime SHIFTINTEGER \n\"Enter a custom shift to the 5\' end rather than the default 2.\"\n"
+    echo -e "  -a, --keepAmbiguousDeletions \n\"Keeps ambiguous deletions in the final output rather than removing them by default.\"\n"
+    echo -e "  -n, --normalizeByBothEnds \n\"Normalize deletions by read depths at the 5\' and 3\' ends rather than just 3\'.\"\n"
+    echo -e "  -h,  --help             \n\"Prints this help\"\n"
+  example
+}
+
+# Ensures that the number of passed args are at least equals
+# to the declared number of mandatory args.
+# It also handles the special case of the -h or --help arg.
+function margs_precheck {
+	if [ $2 ] && [ $1 -lt $margs ]; then
+		if [ $2 == "--help" ] || [ $2 == "-h" ]; then
+			help
+			exit
+		else
+	    	usage
+			example
+	    	exit 1 # error
+		fi
+	fi
+}
+
+# Ensures that all the mandatory args are not empty
+function margs_check {
+	if [ $# -lt $margs ]; then
+	    usage
+	  	example
+	    exit 1 # error
+	fi
+}
+
+# Main method
+#make sure mandatory arguments were input
+margs_precheck $# $1
+
+#set empty variables for mandatory arguments
+fasta=
+testFastq1=
+controlFastq1=
+#set default values for optional arguments
+testFastq2=
+controlFastq2=
+sc=
+nt5Shift=2
+ambig="--NoAmbiguousDeletions"
+norm=
+
+# Args while-loop, parses command line arguments and stores them in above variables
+while [ "$1" != "" ];
+do
+   case $1 in
+   -f  | --fasta )  shift
+                          fasta=$1
+                		  ;;
+   -t1  | --test1 )  shift
+   						  testFastq1=$1
+			              ;;
+   -c1  | --control1  )  shift
+   							controlFastq1=$1
+                          ;;
+   -t2  | --test2  )  shift
+                          testFastq2=$1
+                          ;;
+   -c2  | --control2  )  shift
+                          controlFastq2=$1
+                          ;;
+   -s  | --shift5Prime  )  shift
+                          nt5Shift=$1
+                          ;;
+   -sc  | --structureCassette  )  sc="--sc"
+                          ;;
+   -a  | --keepAmbiguousDeletions  )  ambig=""
+                          ;;  
+   -n  | --normalizeByBothEnds  )  norm="--bothEnds"
+                          ;;                                            
+   -h   | --help )        help
+                          exit
+                          ;;
+   *)                     
+                          echo "$script: illegal option $1"
+                          usage
+						  example
+						  exit 1 # error
+                          ;;
+    esac
+    shift
+done
 
 
 #process reads if they are unparied
-if [ $# -eq 3 ]; then
+if [ ! -e "$testFastq2" ] && [ ! -e "$controlFastq2" ]; then
 	#trim reads
-	mkdir ShapeJumperIntermediateFiles
+	mkdir -p ShapeJumperIntermediateFiles
 	echo "Trimming Reads:"
 	shapemapper_read_trimmer --min_phred 20 --min_length 25 --window_size 5 --in $testFastq1 --out ShapeJumperIntermediateFiles/$testFastq1
 	shapemapper_read_trimmer --min_phred 20 --min_length 25 --window_size 5 --in $controlFastq1 --out ShapeJumperIntermediateFiles/$controlFastq1
 	#copy the fasta file into the intermediate directory
 	cp $fasta ShapeJumperIntermediateFiles
-	#all analysis will take place in this intermeditate directory
+	#all analysis will take place in this intermediate directory
 	cd ShapeJumperIntermediateFiles
 	
 	#Run BWA-MEM to align reads
@@ -69,17 +147,16 @@ if [ $# -eq 3 ]; then
 else
 	#process reads if they are paired
 	#trim reads
-	mkdir ShapeJumperIntermediateFiles
+	mkdir -p ShapeJumperIntermediateFiles
 	echo "Trimming Reads:"
 	shapemapper_read_trimmer --min_phred 20 --min_length 25 --window_size 5 --in $testFastq1 --out ShapeJumperIntermediateFiles/$testFastq1
 	shapemapper_read_trimmer --min_phred 20 --min_length 25 --window_size 5 --in $testFastq2 --out ShapeJumperIntermediateFiles/$testFastq2
 	shapemapper_read_trimmer --min_phred 20 --min_length 25 --window_size 5 --in $controlFastq1 --out ShapeJumperIntermediateFiles/$controlFastq1
 	shapemapper_read_trimmer --min_phred 20 --min_length 25 --window_size 5 --in $controlFastq2 --out ShapeJumperIntermediateFiles/$controlFastq2
 
-	#move original reads into their own directory, move trimmed reads into current directory
 	#copy the fasta file into the intermediate directory
 	cp $fasta ShapeJumperIntermediateFiles
-	#all analysis will take place in this intermeditate directory
+	#all analysis will take place in this intermediate directory
 	cd ShapeJumperIntermediateFiles
 	
 	#Flash the trimmed reads together
@@ -132,27 +209,35 @@ else
 fi
 #find deletions in both samples now that reads have been aligned
 echo "Finding Deletions:"
-python ~/ShapeJumper_V1.0/ShapeJumperScripts/identifyDeletions.py $testSam $fasta -rb -a -e -mi 10 -i 1 -m
-python ~/ShapeJumper_V1.0/ShapeJumperScripts/identifyDeletions.py $controlSam $fasta -rb -a -e -mi 10 -i 1 -m
+python ~/ShapeJumper_V1.0/ShapeJumperScripts/identifyDeletions.py $testSam $fasta -rb $ambig -e -mi 10 
+python ~/ShapeJumper_V1.0/ShapeJumperScripts/identifyDeletions.py $controlSam $fasta -rb $ambig -e -mi 10 
 
 #normalize deletions
 echo "Normalizing Deletions:"
 testDeletions=${testSam:0:${#testSam} - 4}"*_deletions.txt"
 controlDeletions=${controlSam:0:${#controlSam} - 4}"*_deletions.txt"
-python ~/ShapeJumper_V1.0/ShapeJumperScripts/normalizeDeletionRates.py $testDeletions $testSam $fasta --sc
-python ~/ShapeJumper_V1.0/ShapeJumperScripts/normalizeDeletionRates.py $controlDeletions $controlSam $fasta --sc
+python ~/ShapeJumper_V1.0/ShapeJumperScripts/normalizeDeletionRates.py $testDeletions $testSam $fasta $sc $norm
+python ~/ShapeJumper_V1.0/ShapeJumperScripts/normalizeDeletionRates.py $controlDeletions $controlSam $fasta $sc $norm
 
 echo "Subtracting normalized control deletions from normalized crosslinked deletions"
-testNormDeletions=${testSam:0:${#testSam} - 4}"*_normalizedDels.txt"
-controlNormDeletions=${controlSam:0:${#controlSam} - 4}"*_normalizedDels.txt"
+testNormDeletions=${testSam:0:${#testSam} - 4}"*_*ormalizedDels.txt"
+controlNormDeletions=${controlSam:0:${#controlSam} - 4}"*_*ormalizedDels.txt"
 python ~/ShapeJumper_V1.0/ShapeJumperScripts/subtractDeletionRates.py $testNormDeletions $controlNormDeletions --s
 
-echo "Shift 5\' nucleotide position by 2"
+echo "Shift 5\' nucleotide position by $nt5Shift"
 subtractedDeletions=${testNormDeletions:0:${#testNormDeletions} - 4}"_Subtracted.txt"
-python ~/ShapeJumper_V1.0/ShapeJumperScripts/shiftDeletionSites.py $subtractedDeletions --shift5nt 2
+python ~/ShapeJumper_V1.0/ShapeJumperScripts/shiftDeletionSites.py $subtractedDeletions --shift5nt $nt5Shift
 
 #rename shifted deletion file
-shiftedDeletions=${subtractedDeletions:0:${#subtractedDeletions} - 4}"_shift*.txt"
-mv $shiftedDeletions ../${testSam:0:${#testSam} - 4}"_ProcessedDeletions.txt"
+if [ $nt5Shift -eq 0 ]; then
+	shiftedDeletions=${subtractedDeletions}
+else
+	shiftedDeletions=${subtractedDeletions:0:${#subtractedDeletions} - 4}"_shift*.txt"
+fi
+if [ ! -e "$testFastq2" ] && [ ! -e "$controlFastq2" ]; then
+	mv $shiftedDeletions ../${testSam:0:${#testSam} - 4}"_ProcessedDeletions.txt"
+else
+	mv $shiftedDeletions ../${testSam:0:${#testSam} - 11}"_ProcessedDeletions.txt"
+fi
 cd ..
 echo "Done!"
